+++ src/main/java/lotto/LottoGame.java
@@ -1,9 +1,11 @@
 package lotto;
 
 import lotto.domain.LottoNumber;
+import lotto.domain.LottoShop;
 import lotto.domain.Lottos;
 import lotto.domain.PurchaseAmount;
 import lotto.domain.WinningLotto;
+import lotto.domain.generator.RandomLottoGenerator;
 import lotto.view.InputView;
 import lotto.view.ResultView;
