+++ src/main/java/lotto/domain/generator/LottoGenerator.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain.generator;
 
 
 import lotto.domain.Lotto;
