+++ src/main/java/lotto/domain/generator/RandomLottoGenerator.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain.generator;
 
 import lotto.domain.LottoNumber;
 import lotto.domain.Lotto;
