+++ src/main/java/lotto/domain/prize/LottoPrizeMatcher.java
@@ -1,6 +1,4 @@
-package lotto;
-
-import lotto.domain.LottoPrize;
+package lotto.domain.prize;
 
 import java.util.Arrays;
 import java.util.Optional;
