+++ src/test/java/lotto/domain/LottoShopTest.java
@@ -1,8 +1,6 @@
-package lotto;
+package lotto.domain;
 
-import lotto.LottoGeneratorTest.TestLottoGenerator;
-import lotto.domain.Lottos;
-import lotto.domain.PurchaseAmount;
+import lotto.domain.generator.LottoGeneratorTest.TestLottoGenerator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
