+++ src/test/java/lotto/domain/generator/LottoGeneratorTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain.generator;
 
 import lotto.domain.Lotto;
 import lotto.domain.LottoNumber;
