+++ src/test/java/lotto/domain/prize/LottoPrizeMatcherTest.java
@@ -1,6 +1,5 @@
-package lotto;
+package lotto.domain.prize;
 
-import lotto.domain.LottoPrize;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
