+++ src/test/java/lotto/domain/prize/LottoPrizeTest.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.prize;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
